+++ src/it/scala/stellar/sdk/TransactionLedgerEntriesSpec.scala
@@ -2,7 +2,7 @@ package stellar.sdk
 
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.mutable.Specification
-import stellar.sdk.model.ledger.TransactionLedgerEntries
+import stellar.sdk.model.ledger.{LedgerEntryChange, TransactionLedgerEntries}
 import stellar.sdk.model.{Desc, Now}
 
 import scala.concurrent.Await
@@ -13,12 +13,21 @@ class TransactionLedgerEntriesSpec(ee: ExecutionEnv) extends Specification {
 
   import ee.ec
 
+  val last100 = Await.result(PublicNetwork.transactions(cursor = Now, order = Desc), 1.minute).take(100).toList
+
   "transaction meta parsing" should {
     "parse the last 100 without any issues" >> {
       Try {
-        Await.result(PublicNetwork.transactions(cursor = Now, order = Desc), 1.minute).take(100)
-          .map(_.ledgerEntries)
+        last100.map(_.ledgerEntries)
       } must beSuccessfulTry[Seq[TransactionLedgerEntries]]
+    }
+  }
+
+  "fee ledger entry parsing" should {
+    "parse the last 100 without any issues" >> {
+      Try {
+        last100.flatMap(_.feeLedgerEntries)
+      } must beSuccessfulTry[Seq[LedgerEntryChange]]
     }
   }
 }