+++ src/test/scala/stellar/sdk/model/ledger/LedgerEntryGenerators.scala
@@ -2,6 +2,7 @@ package stellar.sdk.model.ledger
 
 import org.scalacheck.{Arbitrary, Gen}
 import stellar.sdk.ArbitraryInput
+import stellar.sdk.model.LedgerThresholds
 import stellar.sdk.model.op.IssuerFlag
 
 trait LedgerEntryGenerators extends ArbitraryInput {
@@ -11,6 +12,11 @@ trait LedgerEntryGenerators extends ArbitraryInput {
     buying <- Gen.posNum[Long]
     selling <- Gen.posNum[Long]
   } yield Liabilities(buying, selling)
+
+  val genLedgerThresholds: Gen[LedgerThresholds] = for {
+    baseThresholds <- genThresholds
+    master <- Gen.choose(0, 255)
+  } yield LedgerThresholds(master, baseThresholds.low, baseThresholds.med, baseThresholds.high)
 
   // LedgerKeys
   val genAccountKey: Gen[AccountKey] = genPublicKey.map(AccountKey.apply)
@@ -37,7 +43,6 @@ trait LedgerEntryGenerators extends ArbitraryInput {
 
   // LedgerEntries
   val genAccountEntry: Gen[AccountEntry] = for {
-    lastModifiedLedgerSeq <- Gen.posNum[Int]
     account <- genPublicKey
     balance <- Gen.posNum[Long]
     seqNum <- Gen.posNum[Long]
@@ -45,39 +50,36 @@ trait LedgerEntryGenerators extends ArbitraryInput {
     inflationDestination <- Gen.option(genPublicKey)
     flags <- Gen.containerOf[Set, IssuerFlag](genIssuerFlag)
     homeDomain <- Gen.option(Gen.identifier)
-    thresholds <- genThresholds
+    thresholds <- genLedgerThresholds
     signers <- Gen.listOf(genSigner)
     liabilities <- Gen.option(genLiabilities)
   } yield AccountEntry(account, balance, seqNum, numSubEntries, inflationDestination, flags, homeDomain,
-    thresholds, signers, liabilities, lastModifiedLedgerSeq)
+    thresholds, signers, liabilities)
 
   val genTrustLineEntry: Gen[TrustLineEntry] = for {
-    lastModifiedLedgerSeq <- Gen.posNum[Int]
     account <- genPublicKey
     asset <- genNonNativeAsset
     balance <- Gen.posNum[Long]
     limit <- Gen.posNum[Long]
     issuerAuthorized <- Gen.oneOf(true, false)
     liabilities <- Gen.option(genLiabilities)
-  } yield TrustLineEntry(account, asset, balance, limit, issuerAuthorized, liabilities, lastModifiedLedgerSeq)
-
-  val genOfferEntry: Gen[OfferEntry] = for {
-    lastModifiedLedgerSeq <- Gen.posNum[Int]
-    account <- genPublicKey
-    offerId <- Gen.posNum[Long]
-    selling <- genAmount
-    buying <- genAsset
-    price <- genPrice
-  } yield OfferEntry(account, offerId, selling, buying, price, lastModifiedLedgerSeq)
+  } yield TrustLineEntry(account, asset, balance, limit, issuerAuthorized, liabilities)
 
   val genDataEntry: Gen[DataEntry] = for {
-    lastModifiedLedgerSeq <- Gen.posNum[Int]
     account <- genPublicKey
     name <- Gen.identifier
     value <- Gen.identifier.map(_.getBytes("UTF-8"))
-  } yield DataEntry(account, name, value, lastModifiedLedgerSeq)
+  } yield DataEntry(account, name, value)
 
-  val genLedgerEntry: Gen[LedgerEntry] = Gen.oneOf(genAccountEntry, genTrustLineEntry, genOfferEntry, genDataEntry)
+  val genLedgerEntryData: Gen[(LedgerEntryData, Int)] = for {
+    idx <- Gen.choose(0, 3)
+    data <- List(genAccountEntry, genTrustLineEntry, genOfferEntry, genDataEntry)(idx)
+  } yield data -> idx
+
+  val genLedgerEntry: Gen[LedgerEntry] = for {
+    lastModifiedLedgerSeq <- Gen.posNum[Int]
+    (data, idx) <- genLedgerEntryData
+  } yield LedgerEntry(lastModifiedLedgerSeq, data, idx)
 
   implicit val arbLedgerEntry = Arbitrary(genLedgerEntry)